rethinking some of the design, adding more tests

+++ src/test/java/com/github/klieber/phantomjs/resolve/PhantomJsBinaryResolverTest.java
@@ -68,7 +68,7 @@ public class PhantomJsBinaryResolverTest {
 
   @AfterClass
   public static void afterClass() {
-    if (phantomJsHome != null) {
+    if (phantomJsHome != null && phantomJsHome.exists()) {
       phantomJsHome.delete();
     }
   }
